requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
requests==2.31.0
python-dotenv==1.0.0